+++ internal/config/config.go
@@ -50,7 +50,7 @@ type BaseOperatorConf struct {
 	UseCustomConfigReloader bool `default:"false"`
 	// container registry name prefix, e.g. docker.io
 	ContainerRegistry                string `default:""`
-	CustomConfigReloaderImage        string `default:"victoriametrics/operator:config-reloader-v0.43.0"`
+	CustomConfigReloaderImage        string `default:"victoriametrics/operator:config-reloader-v0.47.2"`
 	parsedConfigReloaderImageVersion *version.Version
 	PSPAutoCreateEnabled             bool `default:"false"`
 
@@ -73,7 +73,7 @@ type BaseOperatorConf struct {
 
 	VMAlertDefault struct {
 		Image               string `default:"victoriametrics/vmalert"`
-		Version             string `default:"v1.102.0"`
+		Version             string `default:"v1.102.1"`
 		Port                string `default:"8080"`
 		UseDefaultResources bool   `default:"true"`
 		Resource            struct {
@@ -93,7 +93,7 @@ type BaseOperatorConf struct {
 
 	VMAgentDefault struct {
 		Image               string `default:"victoriametrics/vmagent"`
-		Version             string `default:"v1.102.0"`
+		Version             string `default:"v1.102.1"`
 		ConfigReloadImage   string `default:"quay.io/prometheus-operator/prometheus-config-reloader:v0.68.0"`
 		Port                string `default:"8429"`
 		UseDefaultResources bool   `default:"true"`
@@ -113,7 +113,7 @@ type BaseOperatorConf struct {
 
 	VMSingleDefault struct {
 		Image               string `default:"victoriametrics/victoria-metrics"`
-		Version             string `default:"v1.102.0"`
+		Version             string `default:"v1.102.1"`
 		Port                string `default:"8429"`
 		UseDefaultResources bool   `default:"true"`
 		Resource            struct {
@@ -134,7 +134,7 @@ type BaseOperatorConf struct {
 		UseDefaultResources bool `default:"true"`
 		VMSelectDefault     struct {
 			Image    string `default:"victoriametrics/vmselect"`
-			Version  string `default:"v1.102.0-cluster"`
+			Version  string `default:"v1.102.1-cluster"`
 			Port     string `default:"8481"`
 			Resource struct {
 				Limit struct {
@@ -149,7 +149,7 @@ type BaseOperatorConf struct {
 		}
 		VMStorageDefault struct {
 			Image        string `default:"victoriametrics/vmstorage"`
-			Version      string `default:"v1.102.0-cluster"`
+			Version      string `default:"v1.102.1-cluster"`
 			VMInsertPort string `default:"8400"`
 			VMSelectPort string `default:"8401"`
 			Port         string `default:"8482"`
@@ -166,7 +166,7 @@ type BaseOperatorConf struct {
 		}
 		VMInsertDefault struct {
 			Image    string `default:"victoriametrics/vminsert"`
-			Version  string `default:"v1.102.0-cluster"`
+			Version  string `default:"v1.102.1-cluster"`
 			Port     string `default:"8480"`
 			Resource struct {
 				Limit struct {
@@ -204,7 +204,7 @@ type BaseOperatorConf struct {
 	DisableSelfServiceScrapeCreation bool `default:"false"`
 	VMBackup                         struct {
 		Image               string `default:"victoriametrics/vmbackupmanager"`
-		Version             string `default:"v1.102.0-enterprise"`
+		Version             string `default:"v1.102.1-enterprise"`
 		Port                string `default:"8300"`
 		UseDefaultResources bool   `default:"true"`
 		Resource            struct {
@@ -222,7 +222,7 @@ type BaseOperatorConf struct {
 	}
 	VMAuthDefault struct {
 		Image               string `default:"victoriametrics/vmauth"`
-		Version             string `default:"v1.102.0"`
+		Version             string `default:"v1.102.1"`
 		ConfigReloadImage   string `default:"quay.io/prometheus-operator/prometheus-config-reloader:v0.68.0"`
 		Port                string `default:"8427"`
 		UseDefaultResources bool   `default:"true"`